+++ src/components/filter/utils/filterFormUtils.ts
@@ -8,7 +8,7 @@ import { FunctionComponent } from 'react';
 import { FieldConstants } from '../../../utils/constants/fieldConstants';
 import { RangeInput } from '../../inputs/reactHookForm/numbers/RangeInput';
 import { CountriesInput } from '../../inputs/reactHookForm/selectInputs/CountriesInput';
-import { EquipmentType } from '../../../utils/types/equipmentType';
+import { BASE_EQUIPMENTS, EquipmentType } from '../../../utils/types/equipmentType';
 
 const countries = {
     renderer: CountriesInput,
@@ -143,69 +143,10 @@ export const CONTINGENCY_LIST_EQUIPMENTS: Record<
     },
 };
 
-const BASE_EQUIPMENTS: Partial<Record<EquipmentType, { id: EquipmentType; label: string }>> = {
-    [EquipmentType.SUBSTATION]: {
-        id: EquipmentType.SUBSTATION,
-        label: 'Substations',
-    },
-    [EquipmentType.VOLTAGE_LEVEL]: {
-        id: EquipmentType.VOLTAGE_LEVEL,
-        label: 'VoltageLevels',
-    },
-    [EquipmentType.LINE]: {
-        id: EquipmentType.LINE,
-        label: 'Lines',
-    },
-    [EquipmentType.TWO_WINDINGS_TRANSFORMER]: {
-        id: EquipmentType.TWO_WINDINGS_TRANSFORMER,
-        label: 'TwoWindingsTransformers',
-    },
-    [EquipmentType.THREE_WINDINGS_TRANSFORMER]: {
-        id: EquipmentType.THREE_WINDINGS_TRANSFORMER,
-        label: 'ThreeWindingsTransformers',
-    },
-    [EquipmentType.GENERATOR]: {
-        id: EquipmentType.GENERATOR,
-        label: 'Generators',
-    },
-    [EquipmentType.BATTERY]: {
-        id: EquipmentType.BATTERY,
-        label: 'Batteries',
-    },
-    [EquipmentType.LOAD]: {
-        id: EquipmentType.LOAD,
-        label: 'Loads',
-    },
-    [EquipmentType.SHUNT_COMPENSATOR]: {
-        id: EquipmentType.SHUNT_COMPENSATOR,
-        label: 'ShuntCompensators',
-    },
-    [EquipmentType.STATIC_VAR_COMPENSATOR]: {
-        id: EquipmentType.STATIC_VAR_COMPENSATOR,
-        label: 'StaticVarCompensators',
-    },
-    [EquipmentType.DANGLING_LINE]: {
-        id: EquipmentType.DANGLING_LINE,
-        label: 'DanglingLines',
-    },
-};
-
 export const FILTER_EQUIPMENTS: Partial<Record<EquipmentType, { id: EquipmentType; label: string }>> = {
     ...BASE_EQUIPMENTS,
     [EquipmentType.HVDC_LINE]: {
         id: EquipmentType.HVDC_LINE,
         label: 'Hvdc',
-    },
-};
-
-export const SEARCH_EQUIPMENTS: Partial<Record<EquipmentType, { id: EquipmentType; label: string }>> = {
-    ...BASE_EQUIPMENTS,
-    [EquipmentType.HVDC_LINE_LCC]: {
-        id: EquipmentType.HVDC_LINE_LCC,
-        label: 'LCC',
-    },
-    [EquipmentType.HVDC_LINE_VSC]: {
-        id: EquipmentType.HVDC_LINE_VSC,
-        label: 'VSC',
     },
 };
